Speculative fix for bogus PVS warning
#jira UE-150098
#rb none
#preflight 62825666046b81bf93bc31ba

[CL 20225267 by Johan Torp in ue5-main branch]

+++ Engine/Source/Runtime/Core/Private/Serialization/MemoryImage.cpp
@@ -1725,7 +1725,7 @@ FMemoryImageSection* FMemoryImageSection::WritePointer(const FTypeLayoutDesc& St
 	FFrozenMemoryImagePtr FrozenPtr;
 	FrozenPtr.Packed = 0; // dummy value
 	// If derived type matches the static type, store INDEX_NONE to indicate the static type may be used
-	checkf(static_cast<uint64>(TypeDependencyIndex) + 1 <= (1ULL << FFrozenMemoryImagePtr::TypeIndexBits), TEXT(""));
+	checkf(static_cast<uint64>(TypeDependencyIndex) + 1 <= (uint64(1) << FFrozenMemoryImagePtr::TypeIndexBits), TEXT(""));
 	FrozenPtr.SetTypeIndex(bStaticTypeMatchesDerived ? INDEX_NONE : TypeDependencyIndex);
 
 	SectionPointer.Offset = WriteBytes(FrozenPtr);
